string escape: skip NULs in output instead of assrting.
While NULs are not expected in C/ASCII strings, they could be encoded
with unicode:

    $ echo '"a\u0000X"' | ./jq -r '[.]|@tsv' | od -tx1a
    jq: builtin.c:352: escape_string: Assertion `c > 0' failed.

As '\u0000' is a valid unicode symbol. The culprit is
`builtin.c:escape_string()` using `jvp_utf8_next` to iterate the string,
and it returns ascii 0 for '\u0000' - which is valid, but trigger the
assertion.

With this patch, encoded NULLs in the output are silently discarded when
escaping the string with '@tsv' or '@csv':

    $ echo '"a\u0000X"' | ./jq -r '[.]|@tsv' | od -tx1a
    0000000  61  58  0a
              a   X  nl

Note the encoded NULL is kept when not using @tsv/@csv:

    $ echo '"a\u0000X"' | ./jq '[.]'
    [
      "a\u0000X"
    ]