[fixup: main.c]  Fix behaviour in `main` when out of memory (i.e. exit with `2`)
In two other places of `main` in case of `malloc` out-of-memory,
the code calls `perror` and exits with `2`.

In a separate place it uses `fprintf` and exits with `1`.
(Which exit code is reserved for other purposes.)

The patch fixes that one instance by replacing `fprintf` with `perror`
and correctly exits with code `2`.