[fixup: main.c]  In case `jq_init` fails to allocate memory exit immediatly
The previous code on `jq_init` out-of-memory error jumps at the end of `main`,
to the `out` label, which in this case makes no sense.
(For example it would call `jq_teardown` on the `NULL` pointer.)

This patch immediately exits with `2`, just like in other instances in `main`
where an out-of-memory condition is encountered.