jv: Add some support for 64 bit ints in a very conservative way
This adds an extra int64_t field to jv, which is only written when
parsing number literals (and only if they are integers within range),
and only read when printing.

Any operations done over those numbers will downgrade them to a double
with 53 bit precision. This is intentional for the scope of this patch:

    $ echo 111111111111111111 | jq -c '[., 1*.]'
    [111111111111111111,111111111111111100]

For ints between 53 and 64 bits, this matches the behavior of awk,
as suggested in the bug tracker by tischwa:

    $ echo 111111111111111111 | awk '{print $1, 1*$1}'
    111111111111111111 111111111111111104