Fix built-in function contains for strings
This is meant to fix #1732. In the existing code, the libc function
underlying `contains` on strings was `strstr`, which only works
properly on C strings (i.e., arrays of characters, where the first
null is an end marker). This is not suitable for JSON strings, which
can embed null bytes; for example `"xx\u0000yy"` is considered to
include `"\u0000aa"` as a substring, since the latter is interpreted
as the empty string.

This changeset uses `memset` instead of `strstr`.