Add fromstream_with_dups() builtin (fix #1795)
This commit adds a pair of built-ins for dealing with streamed JSON
texts that have duplicate keys.

Eventually we may need a `tostream_with_dups` as well that implements a
convention that array values in objects are to be presented as duplicate
keys with array elements as values.  There are ambiguity issues, like
how should one stream `{"a":[]}`.