jv_unicode: have decoder assume ASCII on invalid first byte
At the start of a UTF-8 sequence, 10xx xxxx is invalid.  Currently, the
decoder falls through to 1111 xxxx and treats this as a 4-byte sequence.  If
a loop is trying to advance some number of UTF-8 codepoints and it hits a
broken byte like this, it is possible to overrun the buffer.  Since the
sequence is invalid anyway, it is likely better to treat it a 1-byte
codepoint.

This has been seen in the wild [1] and if a UTF-8 processor made the
original 4-byte assumption, we can rest assured that it's all invalid anyway
and so preserving the original behavior seems less important that fixing a
buffer overrun.

[1] https://github.com/stedolan/jq/issues/1804