Fix to incorrect --stream paths
Thought something was wrong with the example provided against the --stream option. Proved that the index of the path to "b" was incorrect by running the example on the terminal.  Instead of `[[1,0],"b"]` it should read `[[2,0],"b"]` . 

Additionally, this isn't the whole stream, as given it's depth-first, after visiting descendants, it will visit the ancestor containers too. However I guess these have been omitted deliberately for brevity? 

If you prefer the full sequence I'll force push a fix to this PR branch which contains the ancestors traversed too.  

## Example in terminal

```
echo '[[],"a",["b"]]' | jq --stream --compact-output
[[0],[]]
[[1],"a"]
[[2,0],"b"]
[[2,0]]
[[2]]
```