Fix Autoconf checks on Darwin
The Autoconf checks merely verify that the symbol is available and
linkable, but not that it's actually defined and usable. This breaks
compilation on recent versions of macOS, where `lgamma_r()` is
unavailable unless `_REENTRANT` is defined. Since this is rather
difficult to detect, I strengthened the check so that it simple omits
lgamma_r on macOS.

While at it, I added support for `exp10()` on macOS, which is
available, but named `__exp10` due to its nonstandard nature.