Enable non-single-line mode
Regular expressions are always evaluated in single line mode. This is
because Oniguruma's PERL_NG syntax defaults to single line mode and must
instead be forced _out_ of it by passing ONIG_OPTION_NEGATE_SINGLELINE.
By changing this we now default to non-single line mode, and can opt in
to single line mode as desired.