revamp sub/3 to resolve most issues with gsub (and sub with "g")
The primary purpose of this commit is to rectify most problems with
gsub (and also sub with "g"), and in particular
#1425 (\b), #2354 (lookahead), and #2532 (regex == "^(?!cd ).*$|^cd "; ""))

The revision also partly resolves #2148 and #1206 in that gsub no
longer loops infinitely, but because the new gsub depends on
match(_;"g"), the behavior when regex == "" is sometimes non-standard.

Since the new sub/3 relies on uniq/1, that has been added as well;
this should be a non-issue as the line count of the revised builtin.jq
has been significantly reduced overall.

Also, _nwise/1 has been tweaked to take advantage of TCO.