align documentation of regex with actual behavior; augment onig.test
Several of the FLAGS (as in `test(_; FLAGS)`) simply do not work as advertised, whereas the functionality they promise is available via "Extended Groups".  To make the documentation more useful while allowing for future improvements, it has been revised to focus on what is and ought to be the case.

(Specifically, references to "m" and "s" as possible FLAGS have been dropped, and a concession to reality regarding "p" has been made.)

See Issues #2562 and #2663