regex documentation (FLAGS; Extended Groups) + tests in onig.test
Several of the FLAGS (as in test(_; FLAGS)) simply do not work as advertised, whereas the functionality they promise is available via "Extended Groups". To make the documentation more useful while allowing for future improvements, it has been revised to clarify what is and what ought to be the case.

See Issues #2562 and #2663