Add binary string functionality
This commit adds `tobinary/0`, which converts a normal string (or an
array of bytes, flattened) to a binary string that will be
base64-encoded on output.  The string will remain unencoded during the
execution of the jq program -- encoding is not applied until the string
is to be output, or until `tostring` is applied.

Also added is an `encodeas/1` that takes a string argument of
`"base64"`, `"hex"`, `"bytearray"`, or `"UTF-8"` and outputs its input
altered so that on output the string will be encoded in base64, hex, as
an array of bytes, or be converted to UTF-8 by applying UTF-8 validation
and bad character mapping.

As well, there is a `tobinary/1` that converts a stream of strings,
numeric unsigned byte values, and arrays of bytes, to a binary.

As well there is a `isbinary/0` which indicates whether the input is
binary.

As well there is a `stringtype/0` which indicates whether a string is
binary or UTF-8.

As well there is a `encoding/0` which indicates the output encoding of
the input.